+++ projects/pastanaga-angular/src/lib/chips/chip/chip.component.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { BaseChip } from '../base-chip';
+
+@Component({
+  selector: 'pa-chip',
+  templateUrl: './chip.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class ChipComponent extends BaseChip {}