+++ projects/pastanaga-angular/src/lib/chips/chips.module.ts
@@ -7,10 +7,11 @@ import { PaFocusableModule } from '../focusable';
 import { ChipCloseableComponent } from './chip-closeable/chip-closeable.component';
 import { ChipSelectionableComponent } from './chip-selectionable/chip-selectionable.component';
 import { ChipExpandableComponent } from './chip-expandable/chip-expandable.component';
+import { ChipComponent } from './chip/chip.component';
 
 @NgModule({
   imports: [CommonModule, PaAvatarModule, PaButtonModule, PaIconModule, PaFocusableModule],
-  declarations: [ChipCloseableComponent, ChipSelectionableComponent, ChipExpandableComponent],
-  exports: [ChipCloseableComponent, ChipSelectionableComponent, ChipExpandableComponent],
+  declarations: [ChipComponent, ChipCloseableComponent, ChipSelectionableComponent, ChipExpandableComponent],
+  exports: [ChipComponent, ChipCloseableComponent, ChipSelectionableComponent, ChipExpandableComponent],
 })
 export class PaChipsModule {}