+++ projects/pastanaga-angular/src/lib/popup/popup.directive.ts
@@ -1,8 +1,10 @@
 import {
+  booleanAttribute,
   Directive,
   ElementRef,
   HostListener,
   Input,
+  numberAttribute,
   OnChanges,
   OnDestroy,
   OnInit,
@@ -11,8 +13,6 @@ import {
 } from '@angular/core';
 import { PositionStyle } from '../common';
 import { POPUP_OFFSET, PopupComponent } from './popup.component';
-import { coerceBooleanProperty, coerceNumberProperty } from '@angular/cdk/coercion';
-import { PopupService } from './popup.service';
 import { Subject } from 'rxjs';
 import { takeUntil, throttleTime } from 'rxjs/operators';
 
@@ -23,72 +23,20 @@ import { takeUntil, throttleTime } from 'rxjs/operators';
 export class PopupDirective implements OnInit, OnChanges, OnDestroy {
   @Input() paPopup?: PopupComponent | null;
   @Input() popupPosition?: PositionStyle;
-  @Input()
-  set popupVerticalOffset(value: number) {
-    this._popupVerticalOffset = coerceNumberProperty(value);
-  }
-  get popupVerticalOffset() {
-    return this._popupVerticalOffset;
-  }
-  @Input()
-  get alignPopupOnLeft(): boolean {
-    return this._alignPopupOnLeft;
-  }
-  set alignPopupOnLeft(value: any) {
-    this._alignPopupOnLeft = coerceBooleanProperty(value);
-  }
-  @Input()
-  get popupOnTop(): boolean {
-    return this._popupOnTop;
-  }
-  set popupOnTop(value: any) {
-    this._popupOnTop = coerceBooleanProperty(value);
-  }
-  @Input()
-  get popupOnRight(): boolean {
-    return this._popupOnRight;
-  }
-  set popupOnRight(value: any) {
-    this._popupOnRight = coerceBooleanProperty(value);
-  }
-  @Input()
-  get sameWidth(): boolean {
-    return this._sameWidth;
-  }
-  set sameWidth(value: any) {
-    this._sameWidth = coerceBooleanProperty(value);
-  }
-  @Input()
-  get popupDisabled(): boolean {
-    return this._disabled;
-  }
-  set popupDisabled(value: any) {
-    this._disabled = coerceBooleanProperty(value);
-  }
-  @Input()
-  set openOnly(value: any) {
-    this._openOnly = coerceBooleanProperty(value);
-  }
-  get openOnly() {
-    return this._openOnly;
-  }
+  @Input({ transform: numberAttribute }) popupVerticalOffset = POPUP_OFFSET;
+  @Input({ transform: booleanAttribute }) alignPopupOnLeft = false;
+  @Input({ transform: booleanAttribute }) popupOnTop = false;
+  @Input({ transform: booleanAttribute }) popupOnRight = false;
+  @Input({ transform: booleanAttribute }) sameWidth = false;
+  @Input({ transform: booleanAttribute }) popupDisabled = false;
+  @Input({ transform: booleanAttribute }) openOnly = false;
 
-  private _disabled = false;
-  private _openOnly = false;
-
-  private _alignPopupOnLeft = false;
-  private _popupOnTop = false;
-  private _popupOnRight = false;
-  private _sameWidth = false;
-  private _popupVerticalOffset = POPUP_OFFSET;
   private _handlers: (() => void)[] = [];
-
   private _scrollOrResize = new Subject<Event>();
   private _terminator = new Subject<void>();
 
   constructor(
     private element: ElementRef,
-    private service: PopupService,
     private renderer: Renderer2,
   ) {}
 
@@ -106,10 +54,7 @@ export class PopupDirective implements OnInit, OnChanges, OnDestroy {
   }
 
   ngOnChanges(changes: SimpleChanges) {
-    if (
-      coerceBooleanProperty(changes['popupOnRight']?.currentValue) &&
-      coerceBooleanProperty(changes['alignPopupOnLeft']?.currentValue)
-    ) {
+    if (this.popupOnRight && this.alignPopupOnLeft) {
       console.warn(
         `Incompatible parameters: alignPopupOnLeft and popupOnRight cannot be used at the same time. alignPopupOnLeft is taking precedence.`,
       );
@@ -158,7 +103,7 @@ export class PopupDirective implements OnInit, OnChanges, OnDestroy {
       position: 'fixed',
       top: !this.popupOnTop ? `${top}px` : undefined,
       bottom: this.popupOnTop ? `${bottom}px` : undefined,
-      width: this._sameWidth ? directiveRect.right - directiveRect.left + 'px' : undefined,
+      width: this.sameWidth ? directiveRect.right - directiveRect.left + 'px' : undefined,
     };
 
     const bodyRect = document.body.getBoundingClientRect();