+++ projects/pastanaga-angular/src/lib/table/table-lead-cell-multi-line/table-lead-cell-multi-line.component.ts
@@ -1,5 +1,4 @@
-import { ChangeDetectionStrategy, Component, Input, ViewEncapsulation } from '@angular/core';
-import { coerceBooleanProperty } from '@angular/cdk/coercion';
+import { booleanAttribute, ChangeDetectionStrategy, Component, Input, ViewEncapsulation } from '@angular/core';
 
 @Component({
   selector: 'pa-table-lead-cell-multi-line',
@@ -9,22 +8,6 @@ import { coerceBooleanProperty } from '@angular/cdk/coercion';
   encapsulation: ViewEncapsulation.None,
 })
 export class TableLeadCellMultiLineComponent {
-  @Input()
-  get clickable(): boolean {
-    return this._clickable;
-  }
-  set clickable(value: any) {
-    this._clickable = coerceBooleanProperty(value);
-  }
-
-  @Input()
-  get disabled(): boolean {
-    return this._disabled;
-  }
-  set disabled(value: any) {
-    this._disabled = coerceBooleanProperty(value);
-  }
-
-  private _disabled = false;
-  private _clickable = false;
+  @Input({ transform: booleanAttribute }) clickable = false;
+  @Input({ transform: booleanAttribute }) disabled = false;
 }