+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # 2.63.0 (2023-11-20)
 
-### Improvements
+### Breaking changes
 
 - Upgrade Angular to v17
+- No more `noCloseButton` option on `closeableChip` component. Use basic `pa-chip` component instead.
+- On fields, `showAllErrors` is now false by default
+
+### Improvements
+
 - Use new Angular control flow syntax
+- **Button**: Icon size is now the same as button size by default, and can be overridden with `iconSize` property
 
 # 2.62.7 (2023-11-17)